test(direct-control): extract ready view tests
Move package-owned ready unit and ready city read coverage into focused direct-control package tests with local fake tuner fixtures and no-send assertions.

Record the two framed ready-view handoffs in the support modularization workstream. The parallel agents produced boundary-clean candidates, while the DRA integrated them, removed broad-suite ready-view ownership, and ran package gates in the support worktree.

Validation: bun run --cwd packages/civ7-direct-control test -- ready-unit-view.test.ts ready-city-view.test.ts; bun run --cwd packages/civ7-direct-control test -- ready-unit-view.test.ts ready-city-view.test.ts direct-control.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; git diff --check.
