+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -48,6 +48,10 @@
     thread with explicit context, skills, objective, hard core, exterior,
     falsifier, write set, and return shape. The DRA passively inspected the
     worktree candidate instead of prompting the agent for status.
+  - Direct-control ready-unit and ready-city test lanes used separate fresh
+    `/goal` worktree threads with disjoint one-file write sets. The DRA
+    passively inspected disk candidates and retained broad-suite/docs/commit
+    ownership.
 
 Implementation tasks in sections 3-5 are blocked until the relevant corpus rows
 name the exact write set, fixture owner, validation commands,
@@ -82,14 +86,15 @@ runtime/direct-control claims.
 
 - [x] 4.1 Define direct-control module boundaries and forbidden owners.
 - [ ] 4.2 Add or relocate focused direct-control package tests for each atom.
-- [x] 4.2.1 Add public API/primitives package test coverage.
-- [x] 4.2.2 Add session/framing package test coverage.
-- [x] 4.2.3 Add unit move preview package test coverage.
-- [ ] 4.2.4 Relocate restack-retained broad-suite edge coverage into owning
-  package suites before source extraction: notification panel-dismissal
-  none/expired cases belong to the notification dismissal atom, and log rewrite
-  marker cases belong to the proof/log support atom. They remain in
-  `direct-control.test.ts` only as a restack-speed fallback.
+  - [x] 4.2.1 Add public API/primitives package test coverage.
+  - [x] 4.2.2 Add session/framing package test coverage.
+  - [x] 4.2.3 Add unit move preview package test coverage.
+  - [x] 4.2.4 Add ready unit/city package test coverage.
+  - [ ] 4.2.5 Relocate restack-retained broad-suite edge coverage into owning
+    package suites before source extraction: notification panel-dismissal
+    none/expired cases belong to the notification dismissal atom, and log
+    rewrite marker cases belong to the proof/log support atom. They remain in
+    `direct-control.test.ts` only as a restack-speed fallback.
 - [ ] 4.3 Extract notification view/materialization atom.
 - [ ] 4.4 Extract notification dismissal/verification atom.
 - [ ] 4.5 Extract ready unit/city view atoms.
