+++ openspec/changes/civ7-support-direct-control-modularization/workstream/direct-control-atom-corpus.md
@@ -16,7 +16,7 @@ extraction starts with package-owned atom tests, not source moves.
 | Notification HUD/materialization | Play types around 913-1000, wrapper around 2925, builder around 4714, source around 5119+ | `src/play/notifications/{view,details,decision-hints}.ts` | Package has one broad notification view test; CLI owns the scenario matrix in notification HUD, queue, and priorities suites. | Add `test/play-notification-view.test.ts`; cover package-owned scenario matrix before moving source. | runtime proof for live HUD/source behavior changes | Planned after CLI HUD suite; package proof gap remains. |
 | Notification dismissal/verification | Types around 1005-1045, wrappers around 2935/2946, source around 7491, polling/identity verification around 10472-10534 | `src/play/notifications/{dismissal,verification}.ts` | Package covers verified, engine-front-train-absent, engine-front-dismissed, none-blocker, and expired non-dismissible paths; CLI adds stale nonblocking front evidence. During live-control restack, the none-blocker and expired non-dismissible panel-dismissal tests were retained in the broad suite only as a speed fallback. | Add `test/notification-dismissal.test.ts`; include stale-nonblocking and existing verification modes before source moves. Move the retained panel-dismissal edge tests out of `direct-control.test.ts` in that owner slice. | runtime proof for timing/routes/identity behavior changes | Planned after exact dismiss suite. |
 | Operation validation/send/postconditions | Operation types around 1190-1341, wrappers around 2996-3130, builders 4722/4729, router around 7753, classifiers 10660-11224; diplomacy/narrative wrappers around 3194/3246 | `src/play/operations/{families,validate-request,postconditions,production-choice,chooser-closeouts}.ts` | Package covers unit operation, production choice, unit target basic/no-op, technology closeout, and culture closeout. CLI has delayed/path-shortfall unit-target cases and diplomacy/narrative coverage. | Add focused tests: `unit-operation`, `production-choice`, `unit-target-action`, `technology-choice-closeout`, `culture-choice-closeout`, `diplomacy-response`, and `narrative-choice`; preserve approval-first semantics and no-repeat-after-unverified semantics. | runtime proof for mutation/postcondition behavior changes | Planned after CLI command surfaces stabilize; diplomacy/narrative package proof missing. |
-| Ready unit/city/move preview | Ready-unit types 1392-1449, move preview 1451-1482, ready-city 1484-1549, wrappers 3313/3328/3344, sources 9094/9343/9481 | `src/play/ready/{unit,city,move-preview}.ts` | Package covers ready unit/city and now covers `getCiv7UnitMovePreview` read shape, command routing, bounds rejection, and conservative relationship policy. CLI covers ready unit/city and compact move-preview envelopes. | Add focused `test/ready-unit-view.test.ts` and `test/ready-city-view.test.ts` before source moves; preserve host/port fixture note and relationship-unproven policy shape. | runtime proof for live read/source behavior changes | Unit-move-preview package proof completed; ready unit/city test extraction still pending. |
+| Ready unit/city/move preview | Ready-unit types 1392-1449, move preview 1451-1482, ready-city 1484-1549, wrappers 3313/3328/3344, sources 9094/9343/9481 | `src/play/ready/{unit,city,move-preview}.ts` | Package now covers `getCiv7ReadyUnitView`, `getCiv7ReadyCityView`, and `getCiv7UnitMovePreview` in focused files. Ready-view tests own read-only/no-send assertions and local minimal fixtures; unit-move-preview owns read shape, command routing, bounds rejection, and conservative relationship policy. CLI covers compact envelopes. | Before source moves, preserve public API shape and split source by ready unit, ready city, and move preview owners without creating a broad ready/common catalog. | runtime proof for live read/source behavior changes | Ready package proof completed; source extraction still pending. |
 | Tactical/progression/read lenses | Unit target source 8041, settlement 8258, target candidates 8349, battlefield 8637, destination 8927, traditions 4864, progress dashboard 4975 | `src/play/tactical/*` and `src/play/progression/*` | CLI owns settlement, progression, tactical, and priorities consumer coverage; package lacks direct tests for settlement, target candidates, traditions, progress dashboard, battlefield, and destination reads. | Add `test/settlement-recommendations.test.ts`, `test/traditions-view.test.ts`, `test/progress-dashboard.test.ts`, `test/target-candidates.test.ts`, `test/battlefield-scan.test.ts`, and `test/destination-analysis.test.ts`; assert conservative relationship labels. | read-only live proof only when claiming runtime behavior | Planned; relationship invariant protected. |
 | Map/visibility/gameinfo reads | `index.ts` wrappers around 2358-2469 plus bounded map/grid source regions | `src/play/map/{summary,plot,grid,visibility,gameinfo}.ts` | Package covers map summary, plot snapshot, bounded grid cap, visibility, reveal, and GameInfo rows in broad suite. | Add `test/map-and-visibility.test.ts`; preserve hiddenInfoPolicy and local bounds caps before source moves. | runtime proof only for reveal/mutation behavior changes | Newly added row from package-test boundary report. |
 | Setup and lifecycle orchestration | Setup types around 690-822, wrappers around 2488-2762, setup/start orchestration around 2488-2787 | `src/setup/{snapshot,map-rows,prepare,start,run}.ts` | Package covers setup snapshots/map rows, prepare/start, reload missing map rows, exit-to-shell orchestration, no replay after socket close, seed mismatch, and begin failure. | Add `test/setup-and-lifecycle.test.ts`; preserve no-replay semantics and expected map row proof. | runtime proof for start/setup mutation behavior changes | Newly added row from package-test boundary report. |
