+++ openspec/changes/civ7-support-direct-control-modularization/workstream/review-disposition-ledger.md
@@ -15,4 +15,5 @@
 | Direct-control atom corpus omits public package surfaces and package-test gaps | `019e8af1-0303-7ce2-b059-6178542f833e` | P1 | Accepted/repaired for planning | `direct-control-atom-corpus.md` now names package-test slices before source extraction and adds map/visibility, setup/lifecycle, autoplay/turn/root, and catalog/proof rows that were invisible in the earlier atom table. |
 | Current support branch has no tracked oRPC skill or control-oRPC source | `019e8af1-427e-7463-a9be-dbdeabbccfdf` | P2 | Accepted/deferred | `workstream-record.md`, `design.md`, and `tasks.md` keep oRPC planning-only until `.agents/skills/civ7-orpc-control-architecture` and/or `packages/civ7-control-orpc` authority is imported or explicitly cited from the relevant stack branches. |
 | Session/framing parallel implementation lane must not be treated as proof-complete from the agent worktree | `pendingWorktreeId local:699af531-ba6e-49e6-8f36-021dd3e27721` / disk worktree `6e69` | P1 | Accepted/repaired | The child produced only a boundary-clean candidate file; app thread read/archive tools were unavailable in this session, so DRA passively inspected the disk worktree, integrated and corrected the candidate, removed duplicate broad-suite assertions, and owns final package/OpenSpec gates. |
-| Restack conflict retained new broad-suite edge coverage in a file targeted for modularization | User correction during live-control submit-readiness restack | P2 | Accepted/deferred to owner workstream | `tasks.md` now has 4.2.4: notification panel-dismissal none/expired coverage must move to the notification dismissal atom, and log rewrite marker coverage must move to the proof/log support atom before source extraction. |
+| Ready-view parallel implementation lanes must not be treated as proof-complete from agent worktrees | `pendingWorktreeId local:98a2e225-27e6-4423-abff-c2df614e9547` / disk worktree `ad4e`; `pendingWorktreeId local:31e49457-5f92-4486-b597-624c44e5d18b` / disk worktree `0425` | P1 | Accepted/repaired | The children produced boundary-clean ready-unit and ready-city candidate files only; DRA passively inspected disk worktrees, integrated the candidates, removed duplicate broad-suite assertions and fake branches/helpers, and owns final package/OpenSpec gates. |
+| Restack conflict retained new broad-suite edge coverage in a file targeted for modularization | User correction during live-control submit-readiness restack | P2 | Accepted/deferred to owner workstream | `tasks.md` now has 4.2.5: notification panel-dismissal none/expired coverage must move to the notification dismissal atom, and log rewrite marker coverage must move to the proof/log support atom before source extraction. |
