+++ packages/civ7-direct-control/test/direct-control.test.ts
@@ -556,83 +556,6 @@ describe("Civ7 direct control", () => {
     }
   });
 
-  test("reads the first ready unit view without sending operations", async () => {
-    const server = await startTunerServer();
-    try {
-      const { port } = server.address();
-      const view = await getCiv7ReadyUnitView({}, {
-        host: "127.0.0.1",
-        port,
-        timeoutMs: 1_000,
-      });
-
-      expect(view).toMatchObject({
-        state: { id: "65535", name: "App UI" },
-        unitId: { owner: 0, id: 458752, type: 26 },
-        legalOperations: [
-          {
-            family: "unit-operation",
-            operationType: "SKIP_TURN",
-            valid: true,
-          },
-        ],
-      });
-      expect(server.received.some((message) => message.includes("readReadyUnitView"))).toBe(true);
-      expect(server.received.some((message) => message.includes("sendRequest"))).toBe(false);
-    } finally {
-      await server.close();
-    }
-  });
-
-  test("reads ready-city view for city blockers without sending operations", async () => {
-    const server = await startTunerServer();
-    try {
-      const { port } = server.address();
-      const view = await getCiv7ReadyCityView({}, {
-        host: "127.0.0.1",
-        port,
-        timeoutMs: 1_000,
-      });
-
-      expect(view).toMatchObject({
-        state: { id: "65535", name: "App UI" },
-        localPlayerId: 0,
-        requestedCityId: null,
-        selectedCityId: { ok: true, value: { owner: 0, id: 131073, type: 1 } },
-        cityId: { owner: 0, id: 131073, type: 1 },
-        city: {
-          ok: true,
-          value: {
-            id: { owner: 0, id: 131073, type: 1 },
-            identity: {
-              source: "Players.Cities.getCityIds",
-              ok: true,
-            },
-            name: "Dur-Sharrukin",
-            isTown: true,
-            population: 4,
-          },
-        },
-        legalOperations: [
-          expect.objectContaining({
-            family: "city-operation",
-            operationType: "CONSIDER_TOWN_PROJECT",
-          }),
-        ],
-      });
-      expect(view.notes.some((note) => note.includes("does not choose production"))).toBe(true);
-      expect(view.populationPlacement.ok && view.populationPlacement.value?.cliHints).toContain(
-        "game play expand-city --city-id '<city-id>' --x <x> --y <y>",
-      );
-      expect(server.received.some((message) => message.includes("readReadyCityView"))).toBe(true);
-      expect(server.received.some((message) => message.includes('source: "Players.Cities.getCityIds"'))).toBe(true);
-      expect(server.received.some((message) => message.includes("toComponentId(city.id ?? cityId) ?? cityId"))).toBe(false);
-      expect(server.received.some((message) => message.includes("sendRequest"))).toBe(false);
-    } finally {
-      await server.close();
-    }
-  });
-
   test("classifies shell App UI health when gameplay globals are unavailable", async () => {
     const server = await startTunerServer({
       appUiOnlyStates: true,
@@ -1971,10 +1894,6 @@ async function startTunerServer(options: {
           );
         } else if (frame.message.includes("readUnitTargetAction")) {
           socket.write(encodeResponse(frame.listenerId, [JSON.stringify(unitTargetAction(frame.message.includes('"send":true'), options.unitTargetMode))]));
-        } else if (frame.message.includes("readReadyUnitView")) {
-          socket.write(encodeResponse(frame.listenerId, [JSON.stringify(readyUnitView())]));
-        } else if (frame.message.includes("readReadyCityView")) {
-          socket.write(encodeResponse(frame.listenerId, [JSON.stringify(readyCityView())]));
         } else if (frame.message.includes("sendTechnologyChoiceCloseout")) {
           socket.write(encodeResponse(frame.listenerId, [JSON.stringify(technologyChoiceCloseout())]));
         } else if (frame.message.includes("sendCultureChoiceCloseout")) {
@@ -2505,134 +2424,6 @@ function unitOperationPostconditionSnapshot(firstReadyUnitId: { owner: number; i
     selectedUnitId: { ok: true, value: { owner: 0, id: 65536, type: 26 } },
     firstReadyUnitId: { ok: true, value: firstReadyUnitId },
     blocker: { ok: true, value: 0 },
-  };
-}
-
-function readyUnitView() {
-  const unitId = { owner: 0, id: 458752, type: 26 };
-  return {
-    localPlayerId: 0,
-    requestedUnitId: null,
-    selectedUnitId: { ok: true, value: null },
-    firstReadyUnitId: { ok: true, value: unitId },
-    unitId,
-    unit: {
-      ok: true,
-      value: {
-        id: unitId,
-        owner: 0,
-        type: 111,
-        typeName: "UNIT_ARMY_COMMANDER",
-        location: { x: 22, y: 31 },
-        movementMovesRemaining: 2,
-        attacksRemaining: 0,
-        damage: 0,
-        hitPoints: 100,
-      },
-    },
-    legalOperations: [
-      {
-        family: "unit-operation",
-        operationType: "SKIP_TURN",
-        enumValue: 1,
-        valid: true,
-        result: { Success: true },
-      },
-    ],
-    nearby: {
-      ok: true,
-      value: [
-        {
-          x: 22,
-          y: 31,
-          units: [{ id: unitId, owner: 0, typeName: "UNIT_ARMY_COMMANDER" }],
-        },
-      ],
-    },
-    notes: ["Read-only ready-unit view. Use operation validation before any send."],
-  };
-}
-
-function readyCityView() {
-  const cityId = { owner: 0, id: 131073, type: 1 };
-  return {
-    localPlayerId: 0,
-    requestedCityId: null,
-    selectedCityId: { ok: true, value: cityId },
-    blockingCityId: { ok: true, value: cityId },
-    cityId,
-    city: {
-      ok: true,
-      value: {
-        id: cityId,
-        owner: 0,
-        identity: {
-          source: "Players.Cities.getCityIds",
-          ok: true,
-          observedCityId: cityId,
-          reason: null,
-        },
-        name: "Dur-Sharrukin",
-        location: { x: 22, y: 31 },
-        population: 4,
-        isTown: true,
-        growth: { growthType: -284569333, turnsUntilGrowth: 3 },
-        buildQueue: { currentProductionTypeHash: null, turnsLeft: null },
-      },
-    },
-    legalOperations: [
-      {
-        family: "city-operation",
-        operationType: "CONSIDER_TOWN_PROJECT",
-        enumValue: 1,
-        valid: true,
-        result: { Success: true },
-      },
-    ],
-    productionCandidates: {
-      ok: true,
-      value: [
-        {
-          kind: "constructible",
-          type: 713967338,
-          typeName: "BUILDING_WALLS",
-          name: "LOC_BUILDING_WALLS_NAME",
-          args: { ConstructibleType: 713967338 },
-          valid: true,
-          result: { Success: true, Plots: [1457] },
-          placementPlots: [{ index: 1457, x: 22, y: 31 }],
-          cli: "game play build-production --city-id '<city-id>' --constructible-type 713967338 --x <x> --y <y>",
-        },
-      ],
-    },
-    townFocusOptions: {
-      ok: true,
-      value: [
-        {
-          name: "LOC_PROJECT_FISHING_TOWN_NAME",
-          description: "LOC_PROJECT_FISHING_TOWN_DESCRIPTION",
-          args: { Type: -284569333, ProjectType: -548685232, City: 131073 },
-          valid: true,
-          result: { Success: true },
-          cli: "game play set-town-focus --city-id '<city-id>' --growth-type -284569333 --project-type -548685232",
-        },
-      ],
-    },
-    populationPlacement: {
-      ok: true,
-      value: {
-        isReadyToPlacePopulation: { ok: true, value: true },
-        cityWorkerCap: { ok: true, value: 4 },
-        allPlacementInfo: { ok: true, value: [{ PlotIndex: 1457, IsBlocked: false }] },
-        workablePlotIndexes: { ok: true, value: [1457] },
-        blockedPlotIndexes: { ok: true, value: [] },
-        cliHints: [
-          "game play assign-worker --player-id <id> --location <plot-index>",
-          "game play expand-city --city-id '<city-id>' --x <x> --y <y>",
-        ],
-      },
-    },
-    notes: ["Read-only ready-city view. This view intentionally does not choose production."],
   };
 }
